move app.ts's VOTED logic into helpers in router.ts
